backend/requirements.txt
fastapi
uvicorn
dnspython
aiosmtplib
pandas
python-multipart
email-validator
sqlalchemy
psycopg2-binary
